fix(templates/drizzle-turso): Move Drizzle database setup to index file
