+++ templates/javascript/drizzle-turso-sample/src/function/index.js
@@ -1,6 +1,9 @@
 import { MainRouter } from "../lib/routers/main-router.js";
 import { PageRouter } from "../lib/routers/page-router.js";
 import { ApiRouter } from "../lib/routers/api-router.js";
+import { createClient } from "@libsql/client/web";
+import { drizzle } from 'drizzle-orm/libsql';
+import { integer, text, sqliteTable } from "drizzle-orm/sqlite-core";
 
 async function handleRequest(request, args) {
   // to add or change a route check the route files
@@ -18,7 +21,20 @@ async function handleRequest(request, args) {
     return new Response(`You are trying something incorrect.`, { status: 200 });
   });
 
-  return mainRouter.handle(request, { args });
+  //Setup Turso connection
+  const client = createClient({
+    url: args.url || Azion.env.get("DRIZZLE_TURSO_URL"),
+    authToken: args.token || Azion.env.get("DRIZZLE_TURSO_TOKEN"),
+  });
+  const db = drizzle(client);
+
+  // Setup Turso Table
+  const posts = sqliteTable('posts', {
+    id: integer('id'),
+    message: text('message'),
+  });
+
+  return mainRouter.handle(request, db, posts);
 }
 
 export { handleRequest };
